requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
scikit-learn>=0.24.2
scipy>=1.7.0
matplotlib>=3.4.2
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1 